Update README.md
The code is commented. The code in PythonPart.py is the same as the code of PythonPart.ipynb in PythonPart.zip. To save you time, I changed the suffix of the PythonPart.ipynb file to PythonPart.py to make it convenient for you to view the code directly and save the step of downloading the zip (Due to the size of the PythonPart.ipynb file is over 25MB, so I had to compress it into a ZIP format.).  Tip: You need to install Python's pandas, numpy, plotly, scipy packages before running the code. If you use Visual Studio Code to run the code, please make sure you have installed the Python Third-Party Libraries as follows: From the Menu, click Terminal in the Menu, and then select New Terminal. In the Terminal, you can install the Libraries using pip, which is the Python package management tool. For example, if you want to install the pandas package, you can type the following command: pip install pandas